+++ src/effects/builtin/autopaneffect.h
