+++ src/effects/builtin/balanceeffect.cpp
