+++ src/effects/builtin/bitcrushereffect.h
