+++ src/effects/builtin/builtinbackend.h
@@ -1,0 +1,18 @@
+#ifndef BUILTINBACKEND_H
+#define BUILTINBACKEND_H
+
+#include "effects/effectsbackend.h"
+
+class BuiltInBackend : public EffectsBackend {
+    Q_OBJECT
+  public:
+    BuiltInBackend(QObject* pParent);
+    virtual ~BuiltInBackend();
+
+  private:
+    QString debugString() const {
+        return "BuiltInBackend";
+    }
+};
+
+#endif /* BUILTINBACKEND_H */
