+++ src/effects/builtin/equalizer_util.h
@@ -1,5 +1,5 @@
-#ifndef EFFECTS_NATIVE_EQUALIZER_UTIL_H
-#define EFFECTS_NATIVE_EQUALIZER_UTIL_H
+#ifndef EFFECTS_BUILTIN_EQUALIZER_UTIL_H
+#define EFFECTS_BUILTIN_EQUALIZER_UTIL_H
 
 #include <QObject>
 
@@ -94,4 +94,4 @@ class EqualizerUtil {
 };
 
 
-#endif /* EFFECTS_NATIVE_EQUALIZER_UTIL_H */
+#endif /* EFFECTS_BUILTIN_EQUALIZER_UTIL_H */
