+++ src/effects/builtin/filtereffect.h
