+++ src/effects/builtin/metronomeeffect.cpp
