+++ src/effects/builtin/tremoloeffect.h
