+++ src/effects/effectsmanager.cpp
@@ -70,8 +70,8 @@ bool alphabetizeEffectManifests(EffectManifestPointer pManifest1,
                                 EffectManifestPointer pManifest2) {
     int dNameComp = QString::localeAwareCompare(pManifest1->displayName(), pManifest2->displayName());
     int bNameComp = QString::localeAwareCompare(pManifest1->backendName(), pManifest2->backendName());
-    // Add an exception for "Native" backends, to keep the Native effects in the beginning
-    return (bNameComp ? (bNameComp > 0) : (dNameComp < 0));
+    // Add an exception for "Built-in" backends, to keep the Built-in effects in the beginning
+    return (bNameComp ? (bNameComp < 0) : (dNameComp < 0));
 }
 
 void EffectsManager::addEffectsBackend(EffectsBackend* pBackend) {
