+++ src/effects/lv2/lv2backend.h
@@ -1,9 +1,10 @@
 #ifndef LV2BACKEND_H
 #define LV2BACKEND_H
 
-#include "preferences/usersettings.h"
+#include "effects/defs.h"
 #include "effects/effectsbackend.h"
 #include "effects/lv2/lv2manifest.h"
+#include "preferences/usersettings.h"
 #include <lilv-0/lilv/lilv.h>
 
 class LV2Backend : public EffectsBackend {
